client/requirements.txt
# WebSocket Client (for local server mode)
websockets>=12.0

# HTTP Client (for cloud API mode)
httpx>=0.25.0

# Audio I/O
sounddevice>=0.4.6
numpy>=1.24.0

# Voice Activity Detection
webrtcvad>=2.0.10

# Environment variables
python-dotenv>=1.0.0

# System dependencies (install via apt):
# sudo apt-get install -y portaudio19-dev python3-dev